+++ tests/_Stubs/ReadableFakeEntity.php
@@ -8,10 +8,14 @@ use AppTank\Horus\Illuminate\Database\ReadableEntitySynchronizable;
 class ReadableFakeEntity extends ReadableEntitySynchronizable
 {
 
+    const string ATTR_NAME = "name";
+    const string ATTR_TYPE = "type";
+
     public static function parameters(): array
     {
         return [
-            SyncParameter::createString("name", 1)
+            SyncParameter::createString("name", 1),
+            SyncParameter::createString("type", 1, true)
         ];
     }